Fix code bug with book update when not carried

+++ js/engine.js
@@ -108,7 +108,10 @@ var Engine = (function(global) {
       allEnemies.forEach(function(enemy) {
         if(player.y - enemy.y == 10) {
           if(player.x < enemy.x + 75 && player.x + 75 > enemy.x ){
-            book.drop();
+            // If the player is carrying an item, drop it.
+            if (player.carryItem === true) {
+              book.drop();
+            }
             player.reset();
           }
         }
